Module rust_freertos::kernel

Enums

SchedulerState

Functions

create_idle_task
The fist part of task_start_scheduler(), creates the idle task. Will panic if task creation fails.

task_end_scheduler
NOTE: At the time of writing only the x86 real mode port, which runs on a PC in place of DOS, implements this function.

task_get_scheduler_state

task_increment_tick

task_resume_all
Resumes scheduler activity after it was suspended by a call to vTaskSuspendAll().

task_start_scheduler
Starts the real time kernel tick processing. After calling the kernel has control over which tasks are executed and when.

task_suspend_all
Suspends the scheduler without disabling interrupts. Context switches will not occur while the scheduler is suspended.

task_switch_context
THIS FUNCTION MUST NOT BE USED FROM APPLICATION CODE. IT IS ONLY INTENDED FOR USE WHEN IMPLEMENTING A PORT OF THE SCHEDULER AND IS AN INTERFACE WHICH IS FOR THE EXCLUSIVE USE OF THE SCHEDULER.