Module rust_freertos::list

Structs

List

Descrpition: Definition of the type of queue used by the scheduler.

Implemented by: Fan Jinhao

ListItem

Descrpition: Definition of the only type of object that a list can contain.

Implemented by: Fan Jinhao

Functions

current_list_length

Descrpition: Access macro to return the number of items in the list.

Implemented by: Fan Jinhao

get_list_item_container

Descrpition: Return the list a list item is contained within (referenced from).

Implemented by: Fan Jinhao

get_list_item_owner

Descrpition: Access function to get the owner of a list item. The owner of a list item is the object (usually a TCB) that contains the list item.

Implemented by: Fan Jinhao

get_list_item_value

Descrpition: Access macro to retrieve the value of the list item. The value can represent anything - for example the priority of a task, or the time at which a task should be unblocked.

Implemented by: Fan Jinhao

get_owner_of_head_entry

Descrpition: Access function to obtain the owner of the first entry in a list. Lists are normally sorted in ascending item value order.

get_owner_of_next_entry

Descrpition: Access function to obtain the owner of the next entry in a list.

is_contained_within

Descrpition: Check to see if a list item is within a list. The list item maintains a "container" pointer that points to the list it is in. All this macro does is check to see if the container and the list match.

Implemented by: Fan Jinhao

list_insert

Descrpition: Insert a list item into a list. The item will be inserted into the list in a position determined by its item value (descending item value order).

Implemented by: Fan Jinhao

list_insert_end

Descrpition: Insert a list item into a list. The item will be inserted in a position such that it will be the last item within the list returned by multiple calls to listGET_OWNER_OF_NEXT_ENTRY.

list_is_empty

Descrpition: Access macro to determine if a list contains any items. The macro will only have the value true if the list is empty.

Implemented by: Fan Jinhao

list_remove

Descrpition: Remove an item from a list. The list item has a pointer to the list that it is in, so only the list item need be passed into the function.

Implemented by: Fan Jinhao

set_list_item_owner

Descrpition: Access function to set the owner of a list item. The owner of a list item is the object (usually a TCB) that contains the list item.

Implemented by: Fan Jinhao

set_list_item_value

Descrpition: Access macro to set the value of the list item. In most cases the value is used to sort the list in descending order.

Implemented by: Fan Jinhao

Type Definitions

ItemLink

ListLink

WeakItemLink

WeakListLink