Struct rust_freertos::semaphore::Semaphore

pub struct Semaphore(_);

Methods

impl Semaphore

pub fn new_mutex() -> Self

Descrpition

Create a new mutex type semaphore instance.

Implemented by:Ning Yuting

C implementation:queue.c 504-515

Arguments:

Nothing

Return:

The created mutex.

pub fn get_mutex_holder(&self) -> Option<TaskHandle>

Description

Get the mutex holder.

Implemented by:Lei Siqi

Arguments:

Nothing

Return:

Option<task_control::TaskHandle> - the holder of the mutex

pub fn semaphore_up(&self) -> Result<Option<TaskHandle>, QueueError>

Description

Release a semaphore.

Implemented by:Ning Yuting & Lei Siqi

C implementation:semphr.h 489

Arguments:

Nothing

Return:

Ok(T) if the semaphore was released, otherwise QueueError::QueueEmpty.

pub fn semaphore_down(&self, xBlockTime: TickType) -> Result<(), QueueError>

Description

Obtain a semaphore.

Implemented by:Ning Yuting & Lei Siqi

C implementation:semphr.h 331

Arguments:

xBlockTime - The time in ticks to wait for the semaphore to become available. A block time of zero can be used to poll the semaphore. A block time of portMAX_DELAY can be used to block indefinitely.

Return:

Ok() if the semaphore was obtained, otherwise errQUEUE_FULL.

pub fn create_binary() -> Self

Description

Create a binary semaphore.

Implemented by:Ning Yuting

C implementation:semphr.h 135-144

Arguments:

Nothing

Return:

The created binary semaphore.

pub fn create_counting(max_count: UBaseType) -> Self

Description

Create a counting semaphore.

Implemented by:Ning Yuting

C implementation:semphr.h 1039-1041

Arguments:

max_count - The maximum count value that can be reached. When the semaphore reaches this value it can no longer be 'given'.

Return

The created counting semaphore.

pub fn create_recursive_mutex() -> Self

Description

Created a recursive mutex.

Implemented by:Ning Yuting

C implementation:semphr.h 886-888

Argument

Nothing

Return

The created recursive mutex.

pub fn up_recursive(&self) -> bool

Description

Release a recursive mutex.

Implemented by:Ning Yuting

C implementation:queue.c 570-622

Arguments:

Nothing

Return

bool - true if the recursive mutex was released.

pub fn down_recursive(&self, ticks_to_wait: TickType) -> bool

Description

Obtain a recursive mutex.

Implemented by:Ning Yuting & Lei Siqi

C implementation:queue.c 625-664

Arguments:

ticks_to_wait - The time in ticks to wait for the semaphore to become available. A block time of zero can be used to poll the semaphore.

Return:

bool - true if the recursive mutex was obtained.

pub fn get_recursive_count(&self) -> UBaseType

Description

Get the recursive count of a recursive mutex.

Implemented by:Lei Siqi

Arguments:

Nothing

Return:

UBaseType - the recursive count of the recursive mutex.