Struct rust_freertos::task_control::TaskHandle

pub struct TaskHandle(_);

Type by which tasks are referenced. For example, a call to xTaskCreate returns (via a pointer parameter) an TaskHandle_t variable that can then be used as a parameter to vTaskDelete to delete the task. Since multiple TaskHandles may refer to and own a same TCB at a time, we wrapped TCB within a tuple struct using Arc<RwLock<_>>

Implemented by: Fan Jinhao

Methods

impl TaskHandle

pub fn from_arc(arc: Arc<RwLock<TCB>>) -> Self

pub fn from(tcb: TCB) -> Self

Construct a TaskHandle with a TCB. */

Implemented by: Fan Jinhao.

C implementation:

Arguments

tcb: The TCB that we want to get TaskHandle from.

Return

The created TaskHandle.

pub fn as_raw(self) -> *mut c_void

pub fn get_priority(&self) -> u64

pub fn set_priority(&self, new_priority: u64)

pub fn add_task_to_ready_list(&self) -> Result<(), FreeRtosError>

Place the task represented by pxTCB into the appropriate ready list for the task. It is inserted at the end of the list.

Implemented by: Fan Jinhao.

C implementation:

Arguments

Return

TODO

pub fn get_event_list_item(&self) -> Arc<RwLock<ListItem>>

pub fn get_state_list_item(&self) -> Arc<RwLock<ListItem>>

pub fn get_name(&self) -> String

pub fn get_run_time(&self) -> u32

pub fn set_run_time(&self, next_val: u32) -> u32

pub fn get_delay_aborted(&self) -> bool

pub fn set_delay_aborted(&self, next_val: bool) -> bool

pub fn get_mutex_held_count(&self) -> u64

pub fn set_mutex_held_count(&self, new_count: u64)

pub fn get_base_priority(&self) -> u64

pub fn set_base_priority(&self, new_val: u64)