Struct rust_freertos::task_control::task_control_block

pub struct task_control_block { /* fields omitted */ }

Methods

impl task_control_block

pub fn new() -> Self

pub fn name(self, name: &str) -> Self

pub fn stacksize(self, stacksize: u64) -> Self

pub fn priority(self, priority: u64) -> Self

pub fn initialise<F>(self, func: F) -> Result<TaskHandle, FreeRtosError> where
F: FnOnce() + Send + 'static,

Descrpition:

Internally, within the FreeRTOS implementation, tasks use two blocks of memory. The first block is used to hold the task's data structures. The second block is used by the task as its stack. If a task is created using xTaskCreate() then both blocks of memory are automatically dynamically allocated inside the xTaskCreate() function. (see http://www.freertos.org/a00111.html). If a task is created using xTaskCreateStatic() then the application writer must provide the required memory. xTaskCreateStatic() therefore allows a task to be created without using any dynamic memory allocation.

See xTaskCreateStatic() for a version that does not use any dynamic memory allocation.

xTaskCreate() can only be used to create a task that has unrestricted access to the entire microcontroller memory map. Systems that include MPU support can alternatively create an MPU constrained task using xTaskCreateRestricted().

Implemented by: Fan Jinhao

Arguments:

@param pvTaskCode Pointer to the task entry function. Tasks must be implemented to never return (i.e. continuous loop).

@param pcName A descriptive name for the task. This is mainly used to facilitate debugging. Max length defined by configMAX_TASK_NAME_LEN - default is 16.

@param usStackDepth The size of the task stack specified as the number of variables the stack can hold - not the number of bytes. For example, if the stack is 16 bits wide and usStackDepth is defined as 100, 200 bytes will be allocated for stack storage.

@param pvParameters Pointer that will be used as the parameter for the task being created.

@param uxPriority The priority at which the task should run. Systems that include MPU support can optionally create tasks in a privileged (system) mode by setting bit portPRIVILEGE_BIT of the priority parameter. For example, to create a privileged task at priority 2 the uxPriority parameter should be set to ( 2 | portPRIVILEGE_BIT ).

@param pvCreatedTask Used to pass back a handle by which the created task can be referenced.

Return:

@return pdPASS if the task was successfully created and added to a ready list, otherwise an error code defined in the file projdefs.h

pub fn get_state_list_item(&self) -> Arc<RwLock<ListItem>>

pub fn get_event_list_item(&self) -> Arc<RwLock<ListItem>>

pub fn get_priority(&self) -> u64

pub fn set_priority(&mut self, new_priority: u64)

pub fn get_name(&self) -> String

pub fn get_run_time(&self) -> u32

pub fn set_run_time(&mut self, next_val: u32) -> u32

pub fn get_delay_aborted(&self) -> bool

pub fn set_delay_aborted(&mut self, next_val: bool) -> bool

pub fn get_mutex_held_count(&self) -> u64

pub fn set_mutex_held_count(&mut self, new_count: u64)

pub fn get_base_priority(&self) -> u64

pub fn set_base_priority(&mut self, new_val: u64)